Default input_tf_data.py:batch function now accepts optional arguments to indicate the global logical batch size, and the logical feed process indices.
The trainer.py:SpmdTrainer will dispatch the global physical batch
post-sharding to a logical batch. The minimum global logical batch
size is 1.

This allows one to configure an Input where only a
subset of the processes read the input dataset, and the
remainder are replaced with padding which is discarded by
SpmdTrainer when dispatching the physical batch to logical batch axes.