:sparkles: Allow CloudBuild status lookup across regions (#732)
* ✨ Allow CloudBuild status lookup across regions

It appears the GCloud API for CloudBuild retrieval is region-specific, as
the default is `global` which does not retrieve jobs in `us-central1`
region, etc. An call to get_cloud_build_status() may fail or hang indefinitely
if any CloudBuild jobs are running in a private worker pool in a non-global region.

To mitigate this, cloud_build.py has been updated to include a private
helper function `_get_cloud_build_status_for_region()` which takes a
region as a parameter. The logic for `get_cloud_build_status()` now
iterates through all available regions returned by a private function
_list_available_regions() to find the CloudBuild job in question, then
short-circuiting to avoid unnecessary API calls.

* Dummy commit to kick off build