Correct CAUSAL padding in conv when stride > 1. (#801)
The formula for CAUSAL padding is `(window - stride, stride - 1)`.
Currently, CAUSAL padding uses `(window - 1, 0)`, which only works when the stride is 1.

As a result, the streaming subsampler is currently losing the last frame.
This approach provides a general logic for CAUSAL cases, and after this PR, the
frontend will also reuse this implementation.

For example, with window=15 and stride=6, padding is (9, 5).
Below are examples illustrating how input paddings are transformed into output
paddings across different scenarios.

     left_pad         |         input paddings              -> outputs paddings
 1) |1 1 1|1 1 1|1 1 1|0 0 0|0 0 0|0 0 0|0 0 0|0 0 0|0 0 0| -> 0 0 0
 2) |1 1 1|1 1 1|1 1 1|1 0 0|0 0 0|0 0 0|0 0 0|0 0 0|0 0 0| -> 1 0 0
 3) |1 1 1|1 1 1|1 1 1|1 1 1|1 1 1|0 0 0|0 0 0|0 0 0|0 0 0| -> 1 0 0
 4) |1 1 1|1 1 1|1 1 1|1 1 1|1 1 1|1 0 0|0 0 0|0 0 0|0 0 0| -> 1 1 0
 5) |1 1 1|1 1 1|1 1 1|1 1 1|1 1 1|1 1 1|1 1 1|1 1 1|1 1 1| -> 1 1 1
 6) |1 1 1|1 1 1|1 1 1|0 1 1|1 1 1|1 1 1|1 1 1|1 1 1|1 1 1| -> 0 1 1
 7) |1 1 1|1 1 1|1 1 1|0 0 0|0 0 0|1 1 1|1 1 1|1 1 1|1 1 1| -> 0 1 1
 8) |1 1 1|1 1 1|1 1 1|0 0 0|0 0 0|0 1 1|1 1 1|1 1 1|1 1 1| -> 0 0 1
     |_________________^_________|
                 |_________________^_________|
                             |_________________^_________|
Let's take a closer look at case 7). In case 7), the first window component fully
covers all 0s, so the first component of the output padding should be the last
0 component, meaning the second component is 1.

In case 8), however, the first window component does not cover all 0s, so the
next component should also be 0. If the second component were 1, information
from the last partial window of the input would be lost — which is the current
situation in the code, which this PR wants to fix.

In general, the anchor point should be the next position after the right edge
of the previous window. Since the anchor is defined by the left pad,
`left_pad = window - stride`, and `right_pad = (window - 1) - left_pad`,
simplifying to `right_pad = stride - 1`.

Note: CAUSAL was introduced by me three days ago, and currently, I’m the only
one using it in the streaming encoder. Therefore, logic changes without
backward compatibility are acceptable.