Quantizer does not return one-hot vectors.
onehot can be extremely memory hungry in scenarios with long context and a
large vocab size. For instance, if the vocab size is 256k and the context
length is 32k, it would unnecessarily consume 32GB of memory per batch (256k *
32k * 4 bytes).
Note: ids is 128kB (32k * 4 bytes).

For similar reasons, we avoid using one-hot vectors during codebook lookup and
instead use advanced indexing for memory efficiency.

Note: The change in the unittest is unrelated to any logic changes. The
difference is likely due to not creating temporary tensors in forward(), which
altered the random values of the parameters.