CAUSAL padding=(dilate_window - stride, stride - 1), not (dilate_window - dilate_stride, dilate_stride - 1)
With CAUSAL padding, the lookahead is limited to stride-1 because that portion
becomes part of the past in the next stride.
e.g. window=5, stride=2 -> padding=(3,1)
 0 1 2|3 4 5 6 7 8 9 0 1 2 3 4 5|6
 |_____^_|       <-  ^ is anchor
     |_____^_|
         |_____^_|

Even when dilation > 1, the lookahead remains stride-1. This is because
dilation increases the effective window size, but does not change the stride of
the convolution.
e.g. window=5, stride=2, dilate=2 -> dilate_window=9, padding=(7,1)
 0 1 2 3 4 5 6|7 8 9 0 1 2 3 4 5|6
 |_____________^_|
     |_____________^_|
         |_____________^_|