Allow external positions to be inputed in RoPE embedding layer (#926)
* Allow external positions to be inputed in RoPE embedding layer

Use case: In RoPE embedding, position embeddings are applied to Q, K, V values after `i_proj`. Unlike the implementation of current `RoFormerQKVLinear`, in MaskedDiT we need to customize positions to indicate masked versus non-masked positions in the position embedding. When we convert this masked roformer attention module to flash attention, we need its signature to be supported by `MultiheadAttention`.

* Update attention_test.py

* Update dit.py

* Update attention.py

* Update attention_test.py

* Update attention.py

* Update dit.py

* Update axlearn/common/attention.py

Co-authored-by: Mark Lee <mmaarrkklleeee@gmail.com>

* respond to comments.

Co-authored-by: Ruoming Pang <ruoming@gmail.com>

* Update attention.py

* Update attention.py

* Update attention.py

---------

Co-authored-by: Mark Lee <mmaarrkklleeee@gmail.com>
Co-authored-by: Ruoming Pang <ruoming@gmail.com>