Logmel: use jnp.fft.rfft instead of jnp.fft.fft. (#1221)
Audio samples are always real-valued, which is why the more efficient `rfft`
can be used.

Currently, the logmel frontend uses `fft` and manually slices the output tensor
to `fft_size // 2 + 1`. In contrast, `rfft` directly returns a tensor of shape
`fft_size // 2 + 1` because, when the input is real, half of the FFT output
consists of conjugate complex values, essentially redundant information.